[AIRFLOW-6364] Move conn tests from test_core to test_connection
